requirements/main.txt
anthropic==0.49.0
google-generativeai==0.8.4
    openai==1.66.3
python-dotenv==1.0.1
pillow==11.1.0

# LangChain
langchain==0.3.20
langchain-anthropic==0.3.9
langchain-community==0.3.19
langchain-core==0.3.45
langchain-google-genai==2.0.10
langchain-openai==0.3.8